+++ src/pages/survey/SurveyDetailsPage.tsx
@@ -1,34 +1,10 @@
 import React, { useEffect } from 'react';
 import { useQuery } from 'react-query';
 import { useParams } from 'react-router-dom';
-import apiClient from '../../api/apiClient';
-import User from '../../api/dtos/user.dto';
-import { Reviewer, Survey, Youth } from '../../api/dtos/survey-assignment.dto';
+import { Button, Table, Tbody, Td, Th, Thead, Tr } from '@chakra-ui/react';
+import apiClient, { SurveyDetail, IAssignment } from '../../api/apiClient';
 import LoadingSpinner from '../../components/LoadingSpinner';
 import ErrorAlert from '../../components/ErrorAlert';
-
-export enum AssignmentStatus {
-  INCOMPLETE = 'incomplete',
-  IN_PROGRESS = 'in_progress',
-  COMPLETED = 'complete',
-}
-
-export interface SurveyDetail extends Survey {
-  assignments: IAssignment[];
-}
-
-export interface IAssignment {
-  id: number;
-  uuid: string;
-  survey: Survey;
-  reviewer: Reviewer;
-  youth: Youth;
-  status: AssignmentStatus;
-  sent: boolean;
-  responses: Response[];
-  reminderSent: boolean;
-  started: Date;
-}
 
 const SurveyDetailsPage: React.FC = () => {
   const { survey_uuid: surveyUuid } = useParams<{ survey_uuid: string }>();
@@ -38,6 +14,9 @@ const SurveyDetailsPage: React.FC = () => {
 
   return (
     <>
+      <a href="/private">
+        <Button>Back to surveys</Button>
+      </a>
       {isLoading && <LoadingSpinner />}
       {error && <ErrorAlert />}
       {data && (
@@ -45,32 +24,28 @@ const SurveyDetailsPage: React.FC = () => {
           <h1>
             Survey Details for <span style={{ fontWeight: 'bold' }}>{data.name}</span>
           </h1>
-          <h1>
-            By:{' '}
-            <span style={{ fontWeight: 'bold' }}>
-              {`${data.creator.firstName} ${data.creator.lastName}`}
-            </span>
-          </h1>
-          {data?.assignments.map((assignment: IAssignment) => (
-            <div
-              key={assignment.id}
-              style={{ marginBottom: '20px', padding: '10px', border: '1px solid #ccc' }}
-            >
-              <h2 style={{ fontWeight: 'bold' }}>
-                {' '}
-                {assignment.reviewer.firstName} {assignment.reviewer.lastName}
-              </h2>
-              <p>
-                {' '}
-                <span style={{ fontWeight: 'bold' }}>Youth Name: </span>
-                {`${assignment.youth.firstName} ${assignment.youth.lastName}`}
-              </p>
-              <p>
-                <span style={{ fontWeight: 'bold' }}>Status: </span>
-                {assignment.status}
-              </p>
-            </div>
-          ))}
+          <Table>
+            <Thead>
+              <Tr>
+                <Th>Id</Th>
+                <Th>Reviewer</Th>
+                <Th>Youth</Th>
+                <Th>Status</Th>
+                <Th>Reminder Sent?</Th>
+              </Tr>
+            </Thead>
+            <Tbody>
+              {data.assignments.map((assignment: IAssignment) => (
+                <Tr key={assignment.id}>
+                  <Td>{assignment.id}</Td>
+                  <Td>{`${assignment.reviewer.firstName} ${assignment.reviewer.lastName}`}</Td>
+                  <Td>{`${assignment.youth.firstName} ${assignment.youth.lastName}`}</Td>
+                  <Td>{assignment.status}</Td>
+                  <Td>{assignment.reminderSent ? 'Yes' : 'No'}</Td>
+                </Tr>
+              ))}
+            </Tbody>
+          </Table>
         </>
       )}
     </>
